MBS-9164: Show edit history when not logged in
There's no good reason to completely hide edit history from logged out users.
This makes the edit history visible (including the number of votes) but keeps
both the identity of the voters and all the edit notes hidden. The latter
is a shame, because that's where a lot of the value of the edits being visible
is, but we shouldn't make all edit notes visible without an account since some
users might have posted stuff there that they don't expect to be visible to
everyone outside the MB community (so at the very least we should ask first).