MBS-8931: Invalid ISRC causes an ISE
The TO_JSON method on Server::Form is used to encode values and errors
from FormHandler for display by the JavaScript. Specifically, just the
artist credit editor right now. Unfortunately, the error_fields property
would sometimes contain blessed objects, as was the case with ISRCs. To
fix this, only output the parts of error_fields which we actually need,
and know not to include any blessed objects.

This wasn't a problem for ISWCs because we don't use to_encoded_json on
the work edit page (as stated, it's only needed for artist credits right
now).