Generate cpanfile.snapshot correctly
The carton-local/ directory from an existing image was copied to the
build to speed up the process significantly, so only modules that were
added or changed needed to be installed, and modules that already
satisfied version constraints weren't updated unnecessarily. But, this
didn't handle cases where dependencies were removed from the cpanfile,
and it's arguably more useful to update every module, since we can just
exclude the updates we don't want.