Don't catch errors from $edit->initialize
We rethrow an error regardless, so this isn't letting any additional
ones through.

The first case was intended to pass through certain exception types,
because other code needed to check those. The second case obviously
breaks that by rethrowing a custom/ad-hoc exception with %opts.

We don't really need the %opts anyway, because @_ is visible for each
frame in any stack trace logged to the console. It's sent to Sentry,
too, in addition to the actual request body. If someone needs to deeply
inspect %opts for some reason, they can add their own Data::Dumper
statements.