MBS-8905: Fix allowing to amend release in 1 hour
Previously, allowing to amend release was unreliable, because it was
implemented using `around` modifier on the `allow_auto_edit` method
which was also modified in MusicBrainz::Server::Edit::Release::Edit
using the same modifier, the calling order of which is undefined.

This patch turns the `around` modifier into a regular subroutine.

It also prepares for MBS-8927 to allow amending with conditions.