MBS-8578: Handle NoChanges edits better in the release editor
Prior to this patch, if an edit threw a NoChanges exception it'd cause
the request to 400, and cause it to stop processing edits (a submission
could contain multiple edits). This was a problem, because the release
editor batches recording and medium edits together in one request; at
the same time, the JavaScript ignored 400 errors caused by NoChanges
exceptions (which were indicated by a specific code in the response),
and continued submitting other edits. The result was usually a stub
release with no mediums.

The fix is simply that NoChanges exceptions will no longer cause a
request to /ws/js/edit to fail (though they are still indicated in the
response).