Add user role/predicate for edit search
Prior to this patch, editor and edit note author predicates were
based on various distinct predicates: linked type, ID, and link.

Both editor and edit note author are now based on the same user
role/predicate which will be the common place for further operators.

Incidentally, it fixes a small bug in the edit note author field
that caused the user name to be cleared from the results page.