Rearrange vote count predicate args in edit search
Name argument `vote` so that other arguments are numbered from 0
and still are appended after `vote` to the SQL arguments.

Make a TT macro for the HTML content of predicate `int` field
and share it with predicate Vote count.

Simplify Perl/JS code for Vote countto rely more on ID/int one.