Use UNIX socket to communicate with the renderer
Now that we have a plugin for embedding React components within TT,
reading & writing data to Redis for every call to that plugin has become
prohibitively slow. Use a UNIX socket to communicate with the renderer
instead.

The template-renderer Dockerfile has been removed, and a renderer
service has been added to the existing website container. This is
simpler and probably more correct, since MBS can only communicate with a
renderer on the same host, and it doesn't make sense to communicate with
it as a standalone service, only through the website.