Reduce clearRequireCache calls
The effect of this function is to allow for easier development, so
JavaScript files/templates can be edited without having to restart the
renderer (otherwise, the modules and their associated definitions would
be cached).

This function was previously called at the start of each request to the
renderer. But given that we can now emebed React components within TT,
it's incredibly slow to call this for every component (which corresponds
to what used to be a "request"). So, we now only call this function at
the beginning of each request cycle.