MBS-8640: Make adding work attribute auto-edit
Previously, adding work attribute was not an auto-edit, both because
of a potential deliberate choice and a bug in implementation (`old`
`attributes` array is always defined when adding work attributes).

This patch allows adding work attributes to be an auto-edit, only if
there is no existing attribute of the same type for this work.

It doesn't allow modifying/deleting work attributes to be auto-edit.