Add @RELATABLE_ENTITIES constant
The practical goal of this commit is to reduce the number of calls to
`entities_with` in WebServiceInc, which according to NYTProf is the
top subroutine with the highest exclusive time per request.

Calling `entities_with` with `['mbid', 'relatable']` seems to be common
enough that a constant would be useful anyway.