Remove isrc serialization workaround
ISRCs were being passed to the serializer as an array ref, even though
we only needed the first one (the codes were all the same, only the
recordings differed). This required a workaround in JSON::2::Utils.

We now just store the list of recordings associated with the ISRC code
on the first ISRC entity (in fact, this was already being done, but not
being used), then pass that to the serializer.

The weirdness here is due to the fact that an ISRC *code* can be
associated with many recordings, but an ISRC *entity* (instance) can
only be associated with one. Fixing that would require larger changes to
the data models.