Store last_updated attributes as Str
DateTime->new is quite slow when run on thousands of entities, according
to NYTProf measurements. We don't even really use the last_updated
attribute anywhere that we can't construct a DateTime instance on the
fly. This change knocks between 2-3 seconds off of a 100-area lookup in
the JSON web service (2016 MacBook Pro, 2.6 GHz Intel Core i7).