Pass work attributes via form serialization
The validation in Form::Work had to be changed to prevent the part
labeled 'Convert the value to a format supported by Edit::Work::Edit'
from running if there were any errors. Obviously, that's unnecessary if
there are errors, because no edits will be created. But it's also
problematic because it changes the format to something not understood by
the JavaScript.