MBS-9342: Fix imports with DBD::Pg 3.6.0
DBD::Pg 3.6.0 fixed their UTF8 flag handling with respect to the
pg_(get|put)copydata methods, and the effect is that we now have to
decode our strings before passing them into those functions.

This change also works with DBD::Pg 3.5.3. The way they changed the UTF8
handling is a bit hard to follow since it deals with Perl internals, but
it must have only worked accidentally before, if the behavior we were
relying on wasn't intended.