Fix alias merge logic
Aliases with equivalent values for primary_for_locale and ($type = ?)
will have the same rank; in some cases, this leaves more than one row
with rank = 1 for each partition, and primary_for_locale is not unset
for those additional rows.

This bug was apparently invisible for a long time, because of the
unique_primary_for_locale triggers which enforce uniqueness when all
else fails. Unfortunately, the order in which the UPDATEs inside those
triggers get logged into the dbmirror tables is not correct for some
reason. Edit #45678247 in particular broke replication packet #104949
with artist_alias uniqueness violations on primary_for_locale. See
MBS-9366.