MBS-9372: Allow notes from voting-disabled editor
Prior to this patch, if admin revoked editing/voting privileges of
any editor, this editor would not be able to submit edit notes,
since it would be incidentally disallowed at the same time as votes.

This patch refines permissions to submit edit notes and votes so to
allow submitting edit notes when editor is editing/voting-disabled
and no vote is submitted.  If any vote is submitted at the same
time, edit notes are rejected as well and an error page is returned.