Move sitemap-specific last_modified fallback
JSON dumps will have nullable last_modified columns, since unlike
sitemaps, we aren't required to output these anywhere, so may as well
support indicating that we don't know when the document changed.

Note: These are not to be confused with last_updated columns, which
indicate when the row changed and are not present in these dump schemas.