MBS-9455: Show orderable relationship attributes absent grouping
For orderable link types, we "cleaned" non-required attributes from the
link phrases in `interpolate` so that they are grouped together in the
relationships display and editor UIs. However, the release relationship
editor as well as recording-work relationships on the release page don't
require any grouping, so this had the effect of hiding those attributes
there. This patch makes that behavior optional, so that the cleaned
phrases are only used where grouping is intended.