Upgrade React to v16
We haven't been able to update to any of the intermediate v15 releases,
because we're relying on my fork which adds a <frag> builtin.
Subsequent v15 releases made it very difficult for me to rebase that
work.

React v16 supports fragments (returning an array of elements from
render()) natively, so we no longer need to rely on my fork. We'll keep
the <Frag> idea by defining a wrapper component, because returning an
array is awkward in some instances (unnecessary key props, literals
requiring commas).

The transform-react-inline-elements babel plugin has been removed
because it causes spurious unique key prop warnings. This revealed some
files in which we weren't importing React.