MBS-9467: Handle Patreon user URLs with numeric ID
In Patreon, slug is not mandatory for users which are otherwise
identified by a numeric ID.  This patch allows for both clean URLs
(based on a slug) and fallback user URLs (based on numeric ID).

Silimarly to cleanup for clean URLs, it cleans user URLs up by removing
`/community` and `/posts` parts, superfluous parameters, and hash.