MBS-9513: Read language cookie in BCP 47 format
The `lang` cookie holds locale code in BCP 47 format (with hyphen),
whereas it is compared to POSIX locale codes (with underscore).

Because the `lang` cookie was set in BCP 47 format but read as a
POSIX locale, it caused locales with country code to malfunction.

This patch updates the read of `lang` cookie to follow BCP 47 too.