MBS-9297: Set track lengths far from CD TOC only
Until then, it was possible to set track durations from CD TOC for
any attached medium even if it does not change a millisecond of it.

This patch hides unnecessary links to `/set-durations` and disallows
`SetTrackLengths` medium edits with no change.

Note about hidden pre-gap tracks: these are ignored on purpose,
because CD TOC is not considered reliable regarding their lengths.