Have entity methods not depend on @link_path
This builds up the related entities encountered into %path_ids and fetches
relationships just once at the very end. This is better because we don't end
up fetching the same relationships twice, and we can have our entity methods
skip IDs that were already fetched (which was previously not possible,
because the result of those functions sometimes depended on @link_path).