Replace node-gettext with Jed
While trying to support other domains in our JavaScript, I discovered
from the node-gettext changelog that we'd need to upgrade to v2 in order
to do that. But then I realized there's no technical reason for us to
not simply use Jed on the server (which we already use on the client).
The only technical difference I can see is that we now load the
translations from .po instead of .mo files, but that should have no
effect since they're converted to JS objects either way.

Since Jed is feature complete according to the README, I forked it and
removed the code in it that we don't use or need (sprintf, underscore
implementation). This'll shave some bytes off of our JS bundles,
which'll become more bloated when we start including strings from other
gettext domains.