Add a lens module to replace Immutable.js
Immutable.js doesn't provide much benefit to us, and makes development
harder. It breaks Flow (see #574) and is difficult to type even in the
versions where it works. We have cases where we'd want functions and
components to work with both Immutable instances and plain JS objects
and arrays. That's very cumbersome, since their APIs are different, as
are their typings.

The cases where the Immutable data structures are algorithmically faster
aren't really relevant to the amount of data we display in a single
editing interface. They also aren't necessary to guarantee Immutability.
We can use Flow to create covariant object properties, and use lenses
to perform deep immutable updates on an object. We can also use
Object.freeze if desired.

The lens implementation here was adapted from one by Giulio Canti in the
monocle-ts project (licensed as MIT).