Fix work attribute error display
If an error is set on a subfield (e.g. type_id), it's currently not
displayed, because the JS only looks for errors on the parent compound
field.

subfieldErrors was copied from root/components/forms.js, and modified to
work with plain JS objects. It'll be removed and replaced with a call to
the one from forms.js once we remove Immutable.js.