Fix log message truncation
This was completely broken: adding $nearest_newline to
$MAX_MESSAGE_LENGTH made no sense, because they are both 0-based indexes
into $message. Furthermore, $nearest_newline may very well be at the
*very end* of the original string, defeating the purpose of
$MAX_MESSAGE_LENGTH.

The idea of the original code was, if $MAX_MESSAGE_LENGTH was very close
to a newline character further down the string, we would just include
the whole line so as not to cut off possibly useful data at the end.

Since that was only added because it seemed "nice to have," it's now
removed as "not proven to be useful."