React.embed icons
These need to be moved into JavaScript in order to take advantage of
Webpack's file-loader:
https://webpack.js.org/loaders/file-loader/

This'll allow us to `require()` image files in JavaScript and have them
be copied to our build directory (root/static/build/) and versioned
automatically. At runtime, the require call will return the URL of the
image.