Improve linked_entities serialization
The linked_entities object is used to avoid repeating linked data. For
example, when serializing a list of relationships, we want to send the
link types but not embed (likely duplicating) them in each relationship.

The previous implementation only supported serializing a single entity's
linked entities (into the stash). It was also broken in that it relied
on TO_JSON implementations to invoke the TO_JSON methods on nested
entities directly; otherwise they wouldn't be called until the JSON
encoder was invoked.

That last point hints at the primary issues this commit fixes, which is
that when we encode some arbitrary hash to JSON, everything's already
encoded before we can add the linked_entities to the output. To work
around that, we edit the JSON string directly and splice the
linked_entities in before the closing "}".