Allow setState, including in componentDidMount
While it's preferred to use stateless functional components, some
require state. There's no use in showing warnings in those cases.

Additionally, using setState in componentDidMount is necessary in cases
where the size or position of the mounted DOM node affects the state.
For example, we may have a <Collapsible /> component that hides the part
of its contents that exceed 100px in height.