Remove moment, use Intl.DateTimeFormat
Fixes MBS-10264

moment is a rather bloated library which we barely make use of. And in
fact the places we do use it could easily be using the native Date and
Intl.DateTimeFormat APIs instead.

(To illustrate how large it is, with DEVELOPMENT_SERVER enabled, i.e.
without minification, the size of root/static/build/common-chunks.js is
reduced on my system from 5.0M to 4.3M, just by removing moment and its
extensions.)

Intl.DateTimeFormat works perfectly fine on the server and in modern
browsers. IE <= 11, Edge <= 18, and Safari <= 10 will have issues where
we use formatUserDate in the browser. We fall back to using
Date.prototype.toString() in those rare cases, which I think is
sufficient.

Note: formatUserDate had to be changed to accept $c rather than $c.user
as the first argument, as it now needs access to the current server
language (in the same way that the TT UserDate plugin does).