Unset DBDefs::SEARCH_SERVER during Selenium tests
It currently uses the default, search.musicbrainz.org. So the Selenium
tests are querying our production search.

This causes problems in the (upcoming) release editor test, because the
recording associations tab queries the web service search and overrides
data in t/sql/selenium.sql mapped to the same MBIDs.

It also makes the autocompletes slower. Wherever we have to test the
autocompletes, we have to wait for the indexed search to complete before
we can hit the direct search menu item, which requires a littering of
`pause` commands.

It may make sense to bring this back if we can actually set up a search
server in the test container.