Shore up React <-> knockout bridge
I'm seeing some weird, transient behavior in the release editor's
artist credit editor code while implementing Selenium tests interacting
with it.

I guessed that some of the problems were due to the React components not
re-rendering when certain variables change, so I changed those variables
to knockout observables, and ensured they are subscribed to in a way
that triggers a React render. It seems to prevent the issues I was
seeing.