Add helper utility to hydrate SSR React markup
Hydrating server-side rendered React markup is tricky, because you need
the original props that were used on the server.

Currently, we send the props as a data attribute on some <script> tag,
then use getScriptArgs to retrieve them. This is a bad solution: it
requires the code to have knowledge of which bundle it'll appear in. If
we move to Webpack and start making use of code splitting, this may not
even be a constant we can (or want to) guarantee.

The new solution is to encode the props (as JSON) on the parent element
of the component we need to hydrate. The commons-image implementation
has been updated to use this new method, with one hiccup:

When you hydrate a component, React expects that the markup will remain
identical. We were violating that expectation in commons-image.js by
hydrating using different props (from a web service call).

We now hydrate the CommonsImage component with the same props that were
used on the server, so the markup remains identical. We fetch the actual
image (in case it wasn't cached and provided by the server) in the
componentDidMount lifecycle method.

It may seem odd to make the component in charge of fetching the data
it needs for itself, but that's a direction React is moving toward. See
for example the talk about the upcoming "Suspense" feature here:
https://reactjs.org/blog/2018/03/01/sneak-peek-beyond-react-16.html