Add entity index page bundles
These bundles will hold any code that needs to run on the index page for
an entity. The commons-image.js bundle is removed in favor of explicit
initialization in each entity bundle.

While commons-image.js is an okay solution at present, it would
ultimately become less efficient to continue adding bundles for each
feature rather than each page. For example, we'll soon need to hyrdate
the Wikipedia extract components too; instead of adding a
"wikipedia-extract.js" bundle to mirror the commons-image one, we can
now just add the appropriate intialization code to the bundles for the
entity pages that feature is needed on. We'll eventually want to make
other entity-specific customizations beyond that.

This will also make it much easier to generate bundles containing common
chunks (to borrow Webpack terminology), because we'll be able to detect
exactly what code is shared on each page, rather than having to maintain
that factorization explicitly.