Set x-archive-meta-noindex on CAA uploads
The IA has requested we set this flag on new uploads to prevent them
from getting added to their public search index. They've already purged
previous uploads from the index.

Their reasoning was,

  "We've been digitizing CDs lately and we're finding people somewhat
   confused by whether they're going to get an image or an actual CD
   when looking at search results."

This doesn't affect anything on our side and we'll still be able to
search the entire CAA collection on their side through
https://archive.org/metamgr.php (though it's an admin tool so it
requires the coverartarchive login in syswiki).

As far as the code changes here, they stated we should only need
the noindex tag "once on [our] initial `PUT`" (so I believe this is the
right spot), and that they'd prefer we "use the value `true` [instead
of `1`], just to be consistent with most of the other noindex items on
the site."