Set key for React parameters of l function calls
Localization function `l` (as well as `ln`, `lp`, and `N_l`*) can
take parameters to replace placeholders within the localized string.

If React components are passed as parameter to this function, it
outputs a mix of text and React components which is considered as an
array/list of components by React. Thus, components must have a key.
https://reactjs.org/docs/lists-and-keys.html#keys

This patch set a key to such React components given as parameters to
the function `l`.  Key value is set to the placeholder name as a
style rule (so that we can add parameters without having to check
keys of other parameters).