MBS-9708: Speed up area containment queries
The `area_containment` view has a major flaw, which is that (at least in
PostgreSQL 9.5) the recursive CTE it uses requires calculating
containments for all areas, even though we only need them for the areas
passed to `load_containment` (usually just one area).

We have a cache for the containments set up, but it isn't very effective
because it must be invalidated often (for all areas).

The `area_containment` view was apparently added for SEARCH-359, but it's not
used anywhere in the current or future search indexers.

In this diff, I drop the `area_containment` view and inline it into the
appropriate places, so we can parameterize it and ensure it only
calculates containments for the areas we need.