Add a script to normalize external URLs to HTTPS
This script is to be casually used to update the database after URL
cleanup adds HTTPS normalization for one or more external websites.

Usage: ./admin/NormalizeToHTTPS number_of_rows domain.name...

It is recommended to run it within 10 min, that is, for 500 rows.

URL with authentification (user:password@) are ignored.
URL w/ port number, w/ or w/o path, query or fragment are handled.
Both domains and their subdomains are handled (for the host name).

This commit is motivated by MBS-9646 but applies to many others.
The SQL statement was authored by @chirlu in a comment to MBS-8451.