Interpret usernames as UTF-8 for digest auth
Resolves MBS-8334. Thanks to Ulrich Klauer for his valuable research in
that ticket.

It's not a silver bullet, since clients have to be aware of RFC-7616 or
encode things as UTF-8 in any case. (Note: Picard now uses OAuth.) But
this at least makes digest auth possible for certain usernames. We had
already been rejecting requests whose `Authorization` header could not
be decoded as UTF-8 (as part of the patch for MBS-6360), but had never
been using the decoded username in the actual database lookup.