Add locale-dependent function to hyphenate title
for page/component titles in the form 'Title - Subtitle'.

It is in the script `i18n.js` as hyphenation depends on the locale.

Use it in JS code for release editor.  By the way, it removes two
localizable strings.  Instead, it uses two localizable strings which are
already used in other parts of the code.

Other potential uses of this function are in TT code.  Warning: Make
sure both ends are already used in other parts of the code or update
localizations accordingly, especially for main page title.