Fix React checkboxes
Modify Form::Role::ToJSON to serialize the checkbox value as a boolean.
(Note: Perl treats the string "0" as falsy, so it's handled as false,
as expected, when passed to boolean_to_json.)

Also, browsers tend to use value=on for checkboxes, though our
controllers expect value=1, so change the FormRowCheckbox component to
use that convention instead.