Remove unused /release/lookup endpoint
This page is broken (`medium_sort` tries to access `release->date` so
the page will almost always throw an exception; additionally, the
template references a nonexistent filter), and there are no requests to
it in our gateway logs.

I moved the `DURATION_LOOKUP_RANGE` constant to the WS::2::DiscID
module.