Limit number of releases returned by /ws/2/discid
(This commit is a cherry-pick of 3659d50 from the master branch to
production.)

We occasionally get requests of this form:

  /ws/2/discid/x?toc=y&inc=recordings+artists&limit=1

These requests aren't very common, but a good number of them time out
and generate errors in our logs.

The errors arise due to the number of releases these requests find for
fuzzy TOC lookups, coupled with the use of inc=recordings. The huge
accumulation of tracks attached to these releases is usually enough to
choke the process when passed to load_subobjects or elsewhere. This also
has an adverse effect on memory usage since Perl doesn't release memory
until it exits.

The change I made is to limit the number of releases returned by
/ws/2/discid to 25. This seems reasonable to me, since we order the
fuzzy results by distance. If there isn't a match in the first 25,
you're even less likely to find one later.

I think it's a terrible idea for this endpoint to even return
recordings, but this is less of a breaking change.