Bump Babel to v7.0.0-rc.2
The stage presets were deprecated in favor of specifying the plugins you
actually use.

Since plugins run before presets, but we require the flow preset to run
before the stage presets, the flow preset has also been removed in
favor of specifying the (single) plugin it contains, so we can fix the
order.