Initial port of edit Diff plugin to JavaScript
This is an initial/naïve port of MusicBrainz::Server::Plugin::Diff to
JavaScript/React, minus diff_html_side, which will be removed in a
follow-up commit once the relationship edit diffs are converted.
(Similar to how the release events diffing is done here, we will not
be diffing HTML for relationships either.)

Mostly resolves MBS-10357.