Merge CDStubTOC into CDStub
The basic idea was that CDStubTOC modeled cdtoc_raw, and CDStub modeled
release_raw. But there's no clear reason why these are separate models,
since the tables have a 1:1 mapping (I'm not actually sure why they're
separate tables). You can't submit a CD stub without a disc ID. So these
models only serve to make the code confusing as to when you should
expect which entity.

For example, the CDStub entity class previously had discid and
track_count properties as shortcuts for the indexed search, which didn't
construct a full CDStubTOC in `schema_fixup`. These shortcuts made it
unclear when, for example, $cdstubtoc->discid was defined as opposed to
$cdstubtoc->cdstub->discid, or if $cdstubtoc->cdstub could always be
expected to be loaded.