Move link type serialization into TO_JSON
There shouldn't be more than one method for serializing link types, so
this consolidates them into the standard place (TO_JSON).

Some properties (phrase, reversePhrase, description) are no longer
translated in the serialization, but done so in JavaScript (either on
the client or server). Putting translated strings in the JSON made
caching harder, and we'll likely be moving all translation work to
JavaScript in the future as the website moves to React.

Properties have also been renamed to match their corresponding Moose
attribute names in order to come to a more standardized format and avoid
friction in the codebase.