Split up linkPhrase.interpolate
The previous implementation calculated every combination of results
and returned them in a single array, which is unnecessary in most cases.

It's now split into two functions which can access the phrase or extra
attributes separately. A cache is set up to avoid redundant calculations
between those functions. The cache uses relationship objects as a key,
which will be unique for every page load on the server, but can produce
cache hits on the client (e.g. in the relationship editor).

The caller also now must specify exactly which link phrase to
interpolate, plus whether it will be used for grouping (causing it to
strip non-required attributes from the phrase).