Merge typeInfo and $c.linked_entities
We previously had two ways of accessing linked entities: `typeInfo`
(serialized on the server and passed to the client, to be used in the
relationship and release editors), and `$c.linked_entities` (used on the
server for Node-rendered pages).

This split has become problematic, because we'd like to use some
macros/utility functions that need access to linked entities on both
the client and server. This patch merges these structures together into
a single `linkedEntities` module that is to be accessed on either the
client or server.