Don't mix import with module.exports
ES-module `import` statements and CommonJS `module.exports` assignments
do not work together in the same file with Webpack, as `module.exports`
is set to read-only if the file is determined to be an ES module.

In fact, Babel treats all files as modules by default, so it's important
that we also set the `sourceType` option to "unambiguous" to have it
guess things based on the presence of ES import/export statements.
Otherwise, another issue would arise with the transform-runtime plugin
where it injects `import` statements into files using `module.exports`.