Load Area entities into Place Report
This commit loads area entities into the place report, which can then be accessed by the relevant T4 files with "item.place.area" (assuming that the iterator is called "item").
This was originally made for MBS-9907, so that area data could be implemented into the "PlacesWithoutCoordinates" report within a descriptive link.