Bump React to 16.7.0
Just trying to not lag too far behind before it becomes a pain to
update.

There should be no breaking changes for us. There are a ton of bug
fixes, a few new features, and a couple fixed vulnerabilities.

16.4 did contain a bug fix to getDerivedStateFromProps which caused
issues for people, though I don't think it affected us. Nevertheless, I
removed all of our uses of this method because they turned out to be
unnecessary anti-patterns. See:
https://reactjs.org/blog/2018/06/07/you-probably-dont-need-derived-state.html