MBH-502: SOLR access via search.mb.org/ws/2
We want people with MBS mirrors to continue to be able to use our
production search servers, as they've always been able to do. But
allowing public access to our SOLR cluster is problematic from a
security standpoint. We could set up a proxy [1] in front of SOLR to
block bad requests, but a better option is to just use the /ws/2
compatibility layer already in place in our openresty config. This layer
rewrites "old" API requests understood by our Lucene-based search server
into a format expected by SOLR, and performs an internal redirect.
There's nothing outdated or impractical about this API, so no reason it
needs to be "old."

To allow this, the code for communicating with the old Lucene-based
search is used when SEARCH_ENGINE is 'SOLR' and SEARCH_SERVER is set to
the default public search server (search.musicbrainz.org).

Setting SEARCH_ENGINE to 'LUCENE' is now only necessary when an actual,
old Lucene-based server is configured as SEARCH_SERVER.

[1] https://github.com/dergachev/solr-security-proxy