Replace direct SOLR access with /ws/2 rewrite
We want people with MB mirror servers to continue to be able to use our
search servers, as they've always been able to do. But allowing public
access to our SOLR cluster is problematic from a security standpoint. We
could set up a proxy [1] in front of SOLR to block bad requests, but a
better option is to just use the /ws/2 compatibility layer already in
place in our openresty config. This layer rewrites "old" API requests
understood by our Lucene-based search server into a format expected by
SOLR, and performs an internal redirect. There's nothing outdated or
impractical about this API, so no reason it needs to be "old."

[1] https://github.com/dergachev/solr-security-proxy