Return '' in place of null in expand2
This function should always return a string if no React elements are
present in the result, for compatibility with the old `expand`.

There are likely no use cases where this matters in practice, though
from a logical standpoint I think it makes sense to return an empty
string for '{x:%|}' where x is unset.