Remove the old i18n.expand implementation
expand2 returns plain text, not HTML, so ReactDOMServer's
renderToStaticMarkup has to be used on the result where we actually
expect and need HTML. Since we can't interpolate raw HTML into strings
anymore either, situations where that was done (entity.js) have been
converted to use React elements.

I tried to audit all the places we were passing the result of an l()
call to jQuery's `.html()` setter or knockout's `html:` binding, and
switched them to use `.text()` or `text:`, since the html variants are
dangerous in this case.