Workaround for MBS-9865
This works around the issue while we wait for the release editor to be
converted to React, and for the artist credit editor/bubble code to be
implemented properly.

The main fix is in setting the `target` and `componentInst` data values
on the bubble div before rendering it, *not* after. I don't remember
why that works, exactly, after half an hour in the Chrome debugger,
but it's not important or interesting since this code is going in the
scrap heap.

Fixing this then revealed that the "change matching artists" option was
broken, because someone (me) thought mixing knockout and React was a
good idea to begin with. The solution there is to (handwavy) bring some
codey bits closer to React.