Remove NopArgs check from wrapGettext
We supported translating a NopArgs (which is the return value of N_*
functions) by passing it as the first argument to its corresponding 'l'
function. Since this required you to know which 'l' function it needed,
we later added a toLocaleString method to NopArgs which calls it for
you (by saving the method it needs on the NopArgs instance). We can use
toLocaleString everywhere, then. The check in wrapGettext makes adding
Flow types harder.

Change-Id: I6983bb9021a868e13072fbcf3f13770044c1ea4e